+++ api/src/core/modulos/Resumos/repositories/IResumoRepository.cs
@@ -8,7 +8,7 @@ public interface IResumoRepository {
 
     public Task<Resumo> AtualizarResumo(int Id, Resumo data);
 
-    public Task<Resumo> AtualizarSaldo(string nome, float valor);
+    public Task<Resumo> AtualizarSaldo(string nome, decimal valor);
 
     public Task<List<Resumo>> BuscarTodosResumos();
